{{cookiecutter.github_repository_name}}/requirements_test.txt
django-setuptest
pygments>=2.3.1
pytest>=4.4.1
pytest-django>=3.4.8
pytest-pep8>=1.0.6
pytest-cov
coverage>=4.5.3
pycodestyle
recommonmark
